requirements.txt
# requirements.txt
flask==2.3.3
python-dotenv==1.0.0
werkzeug==2.3.7
gtts==2.5.1
SpeechRecognition==3.10.0
google-generativeai==0.3.2
pillow==10.0.1
numpy==1.24.3
tensorflow==2.13.0
smtplib
email-validator==2.0.0